# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PESkinPro Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PESkinPro E-Commerce Backend",
    "contact": {
      "name": "Developer"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Server (Relative)"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "API Server (Configured)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/posts": {
      "get": {
        "summary": "Ambil Semua Artikel (Blog)",
        "description": "Mengambil daftar semua artikel blog yang sudah dipublikasikan di website.",
        "tags": [
          "Content & Support"
        ],
        "responses": {
          "200": {
            "description": "Daftar artikel berhasil diambil."
          }
        }
      },
      "post": {
        "summary": "Buat Artikel Baru (Admin)",
        "description": "Administrator dapat mempublikasikan artikel blog baru.",
        "tags": [
          "Content & Support"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "description": "Isi artikel dalam format HTML atau string."
                  },
                  "image_url": {
                    "type": "string",
                    "description": "URL gambar utama artikel."
                  },
                  "category_id": {
                    "type": "integer"
                  },
                  "author_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artikel berhasil dibuat."
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Detail Artikel by ID",
        "description": "Mengambil isi lengkap sebuah artikel berdasarkan ID uniknya.",
        "tags": [
          "Content & Support"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data artikel berhasil ditemukan."
          }
        }
      },
      "patch": {
        "summary": "Perbarui Artikel (Admin)",
        "description": "Mengubah isi atau judul artikel yang sudah ada.",
        "tags": [
          "Content & Support"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artikel berhasil diperbarui."
          }
        }
      },
      "delete": {
        "summary": "Hapus Artikel (Admin)",
        "description": "Menghapus artikel blog dari sistem.",
        "tags": [
          "Content & Support"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artikel berhasil dihapus."
          }
        }
      }
    },
    "/posts/slug/{slug}": {
      "get": {
        "summary": "Detail Artikel by Slug",
        "description": "Mengambil isi lengkap sebuah artikel berdasarkan slug (URL-friendly name). Digunakan oleh website storefront.",
        "tags": [
          "Content & Support"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data artikel berhasil ditemukan."
          }
        }
      }
    },
...
    "/cart": {
      "post": {
        "summary": "Tambah ke Keranjang",
        "description": "Menambahkan produk dengan varian tertentu ke dalam keranjang belanja milik pelanggan.",
        "tags": [
          "Order & Checkout"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "product_id",
                  "variant_id"
                ],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "UUID pengguna"
                  },
                  "product_id": {
                    "type": "string"
                  },
                  "variant_id": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Berhasil ditambahkan ke keranjang."
          }
        }
      }
    },
    "/cart/{user_id}": {
      "get": {
        "summary": "Lihat Keranjang",
        "description": "Mengambil semua daftar belanjaan yang ada di dalam keranjang pengguna berdasarkan ID-nya.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data keranjang berhasil diambil."
          }
        }
      }
    },
    "/cart/{id}": {
      "patch": {
        "summary": "Update Jumlah Item",
        "description": "Mengubah jumlah (kuantitas) barang yang ada di dalam keranjang.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID unik item di dalam keranjang."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jumlah barang berhasil diperbarui."
          }
        }
      },
      "delete": {
        "summary": "Hapus dari Keranjang",
        "description": "Mengeluarkan satu jenis barang dari keranjang belanja.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Barang berhasil dihapus dari keranjang."
          }
        }
      }
    },
...
  },
...
}